Moved some code to memory.h and optimized away the header pointer from UnsafeAllocation in release mode.
