+++ Source/DFPSR/base/virtualStack.h
@@ -21,40 +21,12 @@
 //    3. This notice may not be removed or altered from any source
 //    distribution.
 
-
 #ifndef DFPSR_VIRTUAL_STACK
 #define DFPSR_VIRTUAL_STACK
 
 #include "SafePointer.h"
 
 namespace dsr {
-	template <typename T>
-	constexpr uint64_t memory_getPaddedSize() {
-		uint64_t size = (uint64_t)sizeof(T);
-		uint64_t alignment = (uint64_t)alignof(T);
-		// Round up with unsigned integers.
-		return size + (alignment - 1) - ((size - 1) % alignment);
-	}
-
-	// Pre-condition:
-	//   alignment is a power of two (1, 2, 4, 8, 16, 32, 64...)
-	// Post-condition:
-	//   Returns a bit mask for rounding an integer down to the closest multiple of alignment.
-	constexpr uintptr_t memory_createAlignmentAndMask(uintptr_t alignment) {
-		// alignment = ...00001000...
-		// Subtracting one from a power of two gives a mask with ones for the remainder bits.
-		// remainder = ...00000111...
-		// Then we simply negate the mask to get the alignment mask for rounding down.
-		// mask      = ...11111000...
-		return ~(alignment - 1);
-	}
-
-	struct UnsafeAllocation {
-		uint8_t *data;
-		AllocationHeader *header;
-		UnsafeAllocation(uint8_t *data, AllocationHeader *header)
-		: data(data), header(header) {}
-	};
 	// Allocate memory in the virtual stack owned by the current thread.
 	//   paddedSize is the number of bytes to allocate including all elements and internal padding.
 	//     paddedSize must be at least 1, but has no rounding requirements.
@@ -74,7 +46,11 @@ namespace dsr {
 		// Allocate the data with the amount of alignment requested by the element type T.
 		UnsafeAllocation result = virtualStack_push(paddedSize, memory_createAlignmentAndMask((uintptr_t)alignof(T)));
 		// Return a safe pointer to the allocated data.
-		return SafePointer<T>(name, (T*)(result.data), (intptr_t)paddedSize, result.header);
+		#ifdef SAFE_POINTER_CHECKS
+			return SafePointer<T>(name, (T*)(result.data), (intptr_t)paddedSize, result.header);
+		#else
+			return SafePointer<T>(name, (T*)(result.data), (intptr_t)paddedSize);
+		#endif
 	}
 
 	// Free the last allocation from the virtual stack.
